Chapter06/requirements.txt
azureml-sdk[explain,automl]==1.3.0
pandas==0.23.4
numpy==1.16.2
scikit-learn==0.20.3
tensorflow==1.13.2
keras==2.3.1
seaborn==0.10.0
matplotlib==3.2.1
lightgbm==2.3.1